requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==3.1.1
faiss-cpu==1.12.0
PyPDF2==3.0.1
python-multipart==0.0.9
pydantic==1.10.11
torch==2.8.0
streamlit==1.38.0
requests==2.32.3
sentencepiece==0.2.1
typing-extensions>=4.10.0
numpy==1.26.4
transformers==4.44.2
tokenizers==0.19.1